move the FileLock class to Crowbar::Lock
This code is only used by Crowbar, so it makes sense to have it in the
Crowbar namespace.  Also, the use of files is an implementation detail,
so it shouldn't be in the class name in case we decide to switch to
another type of locking in the future.

This commit doesn't change anything except for the class name.
